# Unify the implementation of activation operation