<!-- Demo: https://github.com/PaddlePaddle/Paddle/pull/24810 -->
### PR types
<!-- One of [ New features | Bug fixes | Function optimization | Performance optimization | Breaking changes | Others ] -->
Performance optimization
### PR changes
<!-- One of [ OPs | APIs | Docs | Others ] -->
OPs
### Describe
<!-- Describe what this PR does -->
Unify the implementation of activation operation
本次提交共修改激活算子25个，包括其前向和反向，其中

- 三角函数类的 9 个：sin, cos, tan, asin, acos, atan, sinh, cosh, tanh
- relu 类的 3 个：relu, leaky_relu, ~~elu~~
- sigmoid类的 3 个: sigmoid, silu, logsigmoid
- 舍入类的 3 个：ceil, floor, round
- 数学运算类的 6 个：sqrt, rsqrt, square, log, exp, reciprocal
- 缩放类的 1 个：softshrink

每种类型算子的性能提升近似，因此选取每个类别中的一个算子作为示例进行描述，如下表：
case配置：[16, 128, 257, 257]
|OP Name |FP32 old  |FP32 new  |pro  |FP16 old  |FP16 new  |pro  | 
|---|---|---|---|---|---|---|
|elu fwd |1.6077ms | 1.3114ms| 22.6% |898.68us |670.20us | 34.1% | 
|elu bwd| 2.1628ms| 1.9057ms| 13.5% |1.5737ms |963.07us | 63.4% | 
| sigmoid fwd|1.4083ms| 1.3123ms | 7.3% | 1.0360ms |674.24us  | 53.7% | 
| sigmoid bwd| 2.0002ms | 1.9059ms | 4.9% | 1.1890ms |961.18us  | 23.7% | 
| ceil fwd| 1.5198ms |1.3116ms  | 15.9% | 904.18us |670.92us  | 34.8% | 
| ceil bwd| 1.4069ms|603.45us| 133.1% | 909.00us |302.23us  | 200% | 
| sin fwd| 1.5071ms | 1.3132ms | 14.8% | 989.87us | 673.57us | 47.0% | 
| sin bwd| 2.0647ms | 1.9062ms | 8.3% | 1.3319ms | 970.76us | 37.2% | 
| sqrt fwd| 1.4051ms | 1.3121ms | 7.1% | 950.01us | 672.92us | 41.2% | 
| sqrt bwd| 2.0164ms |1.9069ms  | 5.7% | 1.3418ms | 966.90us | 38.7% | 
| softshrink fwd| 1.5230ms | 1.3118ms | 16.1% | 910.58us | 669.97us | 35.9% | 
| softshrink bwd| 2.0644ms | 1.9057ms | 8.3% | 1.2642ms | 963.07us | 31.3% |